+++ spec/grape/endpoint_spec.rb
@@ -261,6 +261,13 @@ describe Grape::Endpoint do
         end
         post '/omitted_params', MultiJson.encode(:user => 'Blah'), {'CONTENT_TYPE' => 'application/json'}
       end
+
+      it 'should return an equivalent hash on subsequenst calls' do
+        subject.post '/two_times' do
+          body_params.should == body_params
+        end
+        post '/two_times', MultiJson.encode(:user => 'Bobby T.'), {'CONTENT_TYPE' => 'application/json'}
+      end
     end
   end